First implementation of the PositionComponent
Position component is here to replace the legacy UiComponent with a better interface to the ECS.
It should keep all the previous capabilities of the UiComponent but separated as 3 comp: Position, UiAnchor, ClippedTo.
This first allows the position component and it's anchoring to be fully serializable and also reduce drastically the memory footprint of the position component.
This will allow the user to use the position component in the game state without it having all the overhead of the UIComponent while letting any Ui component able to create render call with only the position component
Fixed compilation for linux